+++ TestSumApplication/Models/AddViewModel.cs
@@ -4,9 +4,9 @@ namespace TestSumApplication.Models
 {
     public class AddViewModel
     {
-        [Range(long.MinValue, long.MaxValue, ErrorMessage = "First number must be an valid number")]
-        public long firstNumber { get; set; }
-        [Range(long.MinValue, long.MaxValue, ErrorMessage = "Second number must be an valid number")]
-        public long secondNumber { get; set; }
+        [Range(double.MinValue, double.MaxValue, ErrorMessage = "First number must be an valid number")]
+        public double firstNumber { get; set; }
+        [Range(double.MinValue, double.MaxValue, ErrorMessage = "Second number must be an valid number")]
+        public double secondNumber { get; set; }
     }
 }
